feat: 支持传递 pathMatcheOptions
