+++ packages/core-node/src/index.ts
@@ -1,9 +1,6 @@
 // 输出所有 common 里面会有的内容
 export * from '@opensumi/ide-core-common';
 
-// 输出 node 中特有的内容
-export * from './node-module';
-
 export * from './connection';
 
 export * from './bootstrap';
@@ -11,3 +8,5 @@ export * from './bootstrap';
 export * from './logger/node-logger';
 
 export * from './common-module';
+
+export * from './types';
