+++ packages/extension/src/hosted/ext.process-base.ts
@@ -14,7 +14,7 @@ import {
   ILogService,
 } from '@opensumi/ide-core-common';
 import { isPromiseCanceledError, locale } from '@opensumi/ide-core-common';
-import { AppConfig } from '@opensumi/ide-core-node/lib/bootstrap/app';
+import { AppConfig } from '@opensumi/ide-core-node';
 
 import { ProcessMessageType, IExtensionHostService, KT_PROCESS_SOCK_OPTION_KEY, KT_APP_CONFIG_KEY } from '../common';
 import { CommandHandler } from '../common/vscode';
