+++ packages/extension/src/hosted/extension-log2.ts
@@ -6,7 +6,7 @@ import {
   LogLevel,
   IExtensionLogger,
 } from '@opensumi/ide-core-common';
-import { AppConfig } from '@opensumi/ide-core-node/lib/bootstrap/app';
+import { AppConfig } from '@opensumi/ide-core-node';
 import { LogServiceManager } from '@opensumi/ide-logs/lib/node/log-manager';
 
 export class ExtensionLogger2 implements IExtensionLogger {
